+++ saqc/core/modules/rolling.py
@@ -12,9 +12,12 @@ import numpy as np
 import pandas as pd
 
 from saqc.constants import BAD
+from saqc.lib.docurator import doc
+import saqc.funcs
 
 
 class Rolling:
+    @doc(saqc.funcs.rolling.roll.__doc__)
     def roll(
         self,
         field: str,