+++ CHANGELOG.md
@@ -13,12 +13,17 @@ This changelog starts with version 2.0.0. Basically all parts of the system, inc
 [List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.0.1...develop)
 
 ### Added
-- global keywords documentation ressource
+- global keywords documentation resource added
+- generic documentation module `docurator.py` added to `lib`
+- flagging constants documentation resource added
 ### Changed
+- documentation pipeline changed to base on methods decorators
 - `flagOffsets` parameters `thresh` and `thresh_relative` now both are optional
+- corrected false notion of *residual* concept (old notion: *residue* got replaced by *residual*)
 ### Removed
 ### Fixed
 - `flagOffset` bug with zero-valued threshold
+- `flagCrossStatistics` bug with unaligned input variables
 - `plot` fixed data loss when using *dfilter* kwarg
 
 ## [2.0.1](https://git.ufz.de/rdm-software/saqc/-/tags/v2.0.1) - 2021-12-20